Added initial version of FlexGen web site.
Added home page for the web site. Contains links to the relevant
SourceForge pages, such as the FlexGen project page on SourceForge and
the FlexGen support page. Contains the code specified by SourceForge for
displaying the SourceForge logo and tracking hits on the project web
page. Also contains a basic description of the project.