Adjusted formatting to better handle long tasks.
Tasks with a long enough description would be formatted poorly when
viewed in a web browser. Putting the task description in a seperate div
with the "display" CSS property set to "table" seems to correct the
problem. The previous solution of setting the "display" CSS property of
the marker div to "table" only worked up to task descriptions of a
certain length. This solution seems to work for task descriptions of any
length.