Added another test for MapTile.
Added toString() method to MapUnit to get more useful messages when
tests fail. Added note in the todo file for what to do about the
toString() method if "name" is dropped from MapUnit. Added test ensuring
that the getMapUnit() method returns the correct values for all
coordinates in a small array of map units.