Added mapTileEedges parameter to MapTileType constructor.
Added mapTileEedges parameter to MapTileType constructor. Updated tests
to include the new parameter. Added buildArray() method to
MapTileEdgeHelper to build an array of map tile edges. Updated build
method in MapTileTypeHelper to include a parameter for the size of the
map units array, thus allowing the method to be used in more situations.