Added listener for the event of adding a map tile to a map.
Added MapTileAddedListener interface for defining a listener for the
event of adding a map tile to a map. Updated MapGenerator to support
listeners for the event of adding a map tile to a map. Addded test
ensuring that the addMapTile() method works correctly when a map tile
added listener has been added to the map generator.