Updated the project plan.
Updated the project plan to reflect the state of the current planned
release and future planned releases. Added features to the second
(current) milestone to reflect features that have already been added.
Moved the feature for specifying map features using XML to its own
milestone. Moved the feature for adding support for customizing how the
map is generated to its own milestone. Updated the "todo" list to
reflect the new milestones.